5.partial sums 1 - fixes with data types, add another test

+++ stack_queue_recursion/5.partial_sums_1/main.c
@@ -2,8 +2,8 @@
 #include <stdlib.h>
 #define MAXSIZE 1000010
 
-void precalc(long int *part_sums, const long int *arr, const long int n) {
-    int i;
+void precalc(long long int *part_sums, const long int *arr, const long int n) {
+    long int i;
     part_sums[0] = 0;
     for (i = 1; i <= n; ++i)
         part_sums[i] = part_sums[i-1] + arr[i];
@@ -17,7 +17,7 @@ int main() {
     long int n, Q, a, b, i, j;
 
     long int *arr = malloc(MAXSIZE * sizeof(long int));
-    long int *part_sums = malloc(MAXSIZE * sizeof(long int));
+    long long int *part_sums = malloc(MAXSIZE * sizeof(long int));
 
     fscanf(input, "%li %li", &n, &Q);
     for (i = 1; i <= n; ++i) {
@@ -28,7 +28,7 @@ int main() {
 
     for (j = 0; j < Q; ++j) {
         fscanf(input, "%li %li", &a, &b);
-        fprintf(output, "%li ", part_sums[b]-part_sums[a-1]);
+        fprintf(output, "%lli ", part_sums[b]-part_sums[a-1]);
     }
 
     free(arr);
